Fix deep-audit findings #27-30, #32-33: closes the June 18 audit backlog
- RecursionGuard.Tick (#27): was permanently disabled behind an
  unconditional early-return with a "TEMPORARILY DISABLED... re-enable by
  removing this" comment nobody had acted on. Deliberately NOT re-enabled
  with the old Environment.StackTrace sampling (that's the exact
  allocation it was disabled to rule out) or wired to the sibling
  ThreadStackSampler.SampleAll() (SuspendThreads every thread in the
  process — too invasive to enable unattended). Simplified to an honest,
  documented permanent no-op; ~24 call sites keep compiling unchanged.

- CharacterCaptureHooks.ParseAndSaveCharacterList (#28): the parse loop
  advanced offset (GUID, nameLen, name bytes, padding, timeout) without
  ever comparing it to blobSize — only nameLen itself was range-checked.
  Added bounds checks before each read; breaks (not continues) on any
  overrun since offset is no longer trustworthy past that point.

- AccountHooks (#29): TryGetAccountName/SendNoticeWorldNameDetour validated
  only the start page before a len-driven PtrToStringAnsi copy that could
  cross into an uncommitted page if len were corrupt/torn. Added a
  length-aware IsReadable(ptr, bytes) overload (mirrors CrashLogger's,
  which isn't accessible from this file) plus a len<=256 clamp at all
  three read sites.

- HookResolver.Resolve (#30): on a multi-match, only examined the first
  and immediately-next occurrence, so a 3rd+ match was never considered.
  Not reachable on the current binary (all patterns verified unique) —
  defensive-only for future drift. Now enumerates every match and keeps
  the global-nearest to the fallback VA, with an exact-VA short-circuit.

- SmartBoxLocator.IsMemoryReadable (#32): the region/request-end
  arithmetic used signed 32-bit ToInt32() — acclient.exe is
  LARGE_ADDRESS_AWARE, so addresses above 0x7FFFFFFF are reachable and can
  wrap. Moved the comparison to 64-bit.

- AppraisalHooks (#33): the seven session-scoped per-guid caches had no
  ClearSession() and were absent from DispatchPendingLogout's reset
  pipeline, unlike the peer ObjectQualityCache. Added ClearSession() and
  wired it into PluginManager.DispatchPendingLogout alongside the other
  ResetSession() calls.

This closes every finding in docs/DeepAudit_2026-06-18.md (33/33). All
changes build clean under `dotnet build` and the full NativeAOT
`dotnet publish`.

+++ src/RynthCore.Engine/Compatibility/SmartBoxLocator.cs
@@ -60,9 +60,14 @@ internal static class SmartBoxLocator
         if ((mbi.Protect & READABLE_MASK) == 0)
             return false;
 
-        // Ensure the committed region covers the full requested range
-        int regionEnd = mbi.BaseAddress.ToInt32() + mbi.RegionSize.ToInt32();
-        int requestEnd = address.ToInt32() + size;
+        // Ensure the committed region covers the full requested range.
+        // Deep-audit finding #32 (2026-06-18): this used signed 32-bit
+        // arithmetic (ToInt32()) — acclient.exe is LARGE_ADDRESS_AWARE so
+        // addresses above 0x7FFFFFFF are reachable, and ToInt32() on those
+        // can wrap. Do the comparison in 64-bit (every caller passes a small
+        // compile-time size, so no overflow risk on that side).
+        long regionEnd = mbi.BaseAddress.ToInt64() + mbi.RegionSize.ToInt64();
+        long requestEnd = address.ToInt64() + size;
         return requestEnd <= regionEnd;
     }
